+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [1.4] - 2021-10-13
+### Changed
+- `JSONConfig`: added `configurator` to constructor
+- `JSONConfig`: improved polymorphic deserialization
+
 ## [1.3] - 2021-08-30
 ### Changed
 - `JSONDeserializer`: improved type determination for parameterised types with upper bound